+++ Automation/Core/Configurator.cs
@@ -2,9 +2,6 @@
 // Author: igor.zavoychinskiy@gmail.com
 // License: Public Domain
 
-using Automation.Templates;
-using Automation.Tools;
-using Automation.Utils;
 using Bindito.Core;
 using TimberApi.ConfiguratorSystem;
 using TimberApi.SceneSystem;
@@ -17,12 +14,6 @@ namespace Automation.Core {
 // ReSharper disable once UnusedType.Global
 sealed class Configurator : IConfigurator {
   public void Configure(IContainerDefinition containerDefinition) {
-    CustomToolSystem.BindGroupWithConstructionModeEnabler(containerDefinition, "AutomationToolGroup");
-    CustomToolSystem.BindTool<PauseTool>(containerDefinition);
-    CustomToolSystem.BindTool<ResumeTool>(containerDefinition);
-    CustomToolSystem.BindTool<CancelTool>(containerDefinition);
-    CustomToolSystem.BindTool<ApplyTemplateTool, ApplyTemplateTool.ToolInfo>(containerDefinition);
-    CustomToolSystem.BindTool<DebugPickTool>(containerDefinition);
     containerDefinition.MultiBind<TemplateModule>().ToProvider(ProvideTemplateModule).AsSingleton();
     containerDefinition.Bind<AutomationService>().AsSingleton();
   }