+++ Automation/package/CHANGES.md
@@ -1,3 +1,9 @@
+# v0.13 (18 Feb 2024):
+* Add "finish now" debug tool to immediatelty complete construction of the selected objects.
+* Fix detecting current drought state condition.
+* Change weather icons to the current game scheme.
+* Add new skins for the debug tools.
+
 # v0.12 (10 Jan 2024)
 * Experimental version for Update 5.
 * [Fix #22] Mod doesn't load in Timberborn 0.5.6.3.